Code Refactoring, Added Timeout Function.
