+++ src/index.html
@@ -1,4 +1,10 @@
 <html>
     <head></head>
-    <body></body>
+    <body>
+        <div id="rootPlaceholder"></div>
+        <div>
+            <p>Catch Timer: <span id="catchTimer"></span></p>
+            <p>Catch Count: <span id="catchCount"></span></p>
+        </div>
+    </body>
 </html>
